+++ MarfatechRabbitQueueBundle.php
@@ -2,10 +2,25 @@
 
 declare(strict_types=1);
 
+/*
+ * This file is part of the RabbitQueueBundle package.
+ *
+ * (c) MarfaTech <https://marfa-tech.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace MarfaTech\Bundle\RabbitQueueBundle;
 
+use MarfaTech\Bundle\RabbitQueueBundle\DependencyInjection\Compiler\AMQPStreamConnectionPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class MarfatechRabbitQueueBundle extends Bundle
 {
+    public function build(ContainerBuilder $container): void
+    {
+        $container->addCompilerPass(new AMQPStreamConnectionPass());
+    }
 }